# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Parvath FinServ CRM",
    "version": "1.0.0",
    "description": "All business endpoints require a workspace-scoped session. Mutations require the CSRF token returned by /auth/csrf or /auth/login. Dates use Asia/Kolkata; money is an integer string in paise. See docs/api.md for lifecycle contracts."
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "parvath.sid"
      }
    },
    "schemas": {
      "Client": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 120
          },
          "phone": {
            "type": "string"
          },
          "email": {
            "anyOf": [
              {
                "anyOf": [
                  {
                    "type": "string",
                    "format": "email",
                    "pattern": "^(?:[A-Za-z0-9_'+\\-]+\\.)*[A-Za-z0-9_'+\\-]*[A-Za-z0-9_+-]@(?:[A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
                  },
                  {
                    "type": "string",
                    "const": ""
                  }
                ]
              },
              {
                "type": "null"
              }
            ]
          },
          "kind": {
            "default": "Individual",
            "type": "string",
            "enum": [
              "Individual",
              "Business"
            ]
          },
          "dob": {
            "anyOf": [
              {
                "anyOf": [
                  {
                    "type": "string",
                    "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                  },
                  {
                    "type": "string",
                    "const": ""
                  }
                ]
              },
              {
                "type": "null"
              }
            ]
          },
          "gender": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ]
          },
          "occupation": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ]
          },
          "address": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ]
          },
          "city": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ]
          },
          "state": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ]
          },
          "source": {
            "default": "Direct",
            "type": "string",
            "maxLength": 80
          },
          "annualIncome": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ]
          },
          "riskProfile": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ]
          },
          "investmentInterest": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ]
          },
          "loanInterest": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ]
          },
          "preferredContact": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ]
          },
          "notesText": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ]
          },
          "onboardingProfile": {
            "type": "object",
            "properties": {
              "maritalStatus": {
                "type": "string",
                "maxLength": 1000
              },
              "spouseName": {
                "type": "string",
                "maxLength": 1000
              },
              "spouseDob": {
                "anyOf": [
                  {
                    "type": "string",
                    "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                  },
                  {
                    "type": "string",
                    "const": ""
                  }
                ]
              },
              "weddingDate": {
                "anyOf": [
                  {
                    "type": "string",
                    "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                  },
                  {
                    "type": "string",
                    "const": ""
                  }
                ]
              },
              "children": {
                "maxItems": 20,
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string",
                      "maxLength": 1000
                    },
                    "dob": {
                      "anyOf": [
                        {
                          "type": "string",
                          "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                        },
                        {
                          "type": "string",
                          "const": ""
                        }
                      ]
                    },
                    "relationship": {
                      "type": "string",
                      "maxLength": 1000
                    }
                  }
                }
              },
              "dependents": {
                "type": "string",
                "maxLength": 1000
              },
              "sameAddress": {
                "type": "boolean"
              },
              "permanentAddress": {
                "type": "string",
                "maxLength": 1000
              },
              "pinCode": {
                "type": "string",
                "maxLength": 1000
              },
              "companyName": {
                "type": "string",
                "maxLength": 1000
              },
              "designation": {
                "type": "string",
                "maxLength": 1000
              },
              "professionalIndustry": {
                "type": "string",
                "maxLength": 1000
              },
              "experience": {
                "type": "string",
                "maxLength": 1000
              },
              "referredBy": {
                "type": "string",
                "maxLength": 1000
              },
              "clientCategory": {
                "type": "string",
                "maxLength": 1000
              },
              "clientSince": {
                "anyOf": [
                  {
                    "type": "string",
                    "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                  },
                  {
                    "type": "string",
                    "const": ""
                  }
                ]
              },
              "monthlySavings": {
                "type": "string",
                "maxLength": 1000
              },
              "totalSavings": {
                "type": "string",
                "maxLength": 1000
              },
              "investmentHorizon": {
                "type": "string",
                "maxLength": 1000
              },
              "financialGoals": {
                "maxItems": 20,
                "type": "array",
                "items": {
                  "type": "string",
                  "maxLength": 80
                }
              },
              "policies": {
                "maxItems": 30,
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string",
                      "maxLength": 1000
                    },
                    "provider": {
                      "type": "string",
                      "maxLength": 1000
                    },
                    "name": {
                      "type": "string",
                      "maxLength": 1000
                    },
                    "sumAssured": {
                      "type": "string",
                      "maxLength": 1000
                    },
                    "renewalDate": {
                      "anyOf": [
                        {
                          "type": "string",
                          "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                        },
                        {
                          "type": "string",
                          "const": ""
                        }
                      ]
                    }
                  }
                }
              },
              "loans": {
                "maxItems": 30,
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string",
                      "maxLength": 1000
                    },
                    "bank": {
                      "type": "string",
                      "maxLength": 1000
                    },
                    "amount": {
                      "type": "string",
                      "maxLength": 1000
                    },
                    "outstanding": {
                      "type": "string",
                      "maxLength": 1000
                    },
                    "emi": {
                      "type": "string",
                      "maxLength": 1000
                    },
                    "closureDate": {
                      "anyOf": [
                        {
                          "type": "string",
                          "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                        },
                        {
                          "type": "string",
                          "const": ""
                        }
                      ]
                    }
                  }
                }
              },
              "investments": {
                "maxItems": 30,
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string",
                      "maxLength": 1000
                    },
                    "provider": {
                      "type": "string",
                      "maxLength": 1000
                    },
                    "amount": {
                      "type": "string",
                      "maxLength": 1000
                    },
                    "maturityDate": {
                      "anyOf": [
                        {
                          "type": "string",
                          "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                        },
                        {
                          "type": "string",
                          "const": ""
                        }
                      ]
                    },
                    "expectedReturn": {
                      "type": "string",
                      "maxLength": 1000
                    }
                  }
                }
              },
              "communicationChannels": {
                "maxItems": 10,
                "type": "array",
                "items": {
                  "type": "string",
                  "maxLength": 40
                }
              },
              "preferredTime": {
                "type": "string",
                "maxLength": 1000
              },
              "bestDay": {
                "type": "string",
                "maxLength": 1000
              },
              "engagementTopics": {
                "maxItems": 20,
                "type": "array",
                "items": {
                  "type": "string",
                  "maxLength": 80
                }
              },
              "interests": {
                "maxItems": 20,
                "type": "array",
                "items": {
                  "type": "string",
                  "maxLength": 80
                }
              },
              "importantDates": {
                "maxItems": 30,
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string",
                      "maxLength": 1000
                    },
                    "date": {
                      "anyOf": [
                        {
                          "type": "string",
                          "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                        },
                        {
                          "type": "string",
                          "const": ""
                        }
                      ]
                    },
                    "description": {
                      "type": "string",
                      "maxLength": 1000
                    }
                  }
                }
              },
              "language": {
                "type": "string",
                "maxLength": 1000
              },
              "noCallsDuringHours": {
                "type": "boolean"
              },
              "whatsappOnly": {
                "type": "boolean"
              },
              "initialFollowup": {
                "type": "object",
                "properties": {
                  "enabled": {
                    "type": "boolean"
                  },
                  "date": {
                    "anyOf": [
                      {
                        "type": "string",
                        "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                      },
                      {
                        "type": "string",
                        "const": ""
                      }
                    ]
                  },
                  "channel": {
                    "type": "string",
                    "enum": [
                      "Call",
                      "WhatsApp",
                      "Email",
                      "Meeting"
                    ]
                  },
                  "notes": {
                    "type": "string",
                    "maxLength": 1000
                  }
                },
                "required": [
                  "enabled"
                ]
              }
            }
          },
          "registrationNumber": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ]
          },
          "industry": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ]
          },
          "allowDuplicate": {
            "default": false,
            "type": "boolean"
          },
          "duplicateReason": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ]
          },
          "version": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "maximum": 9007199254740991
          },
          "tags": {
            "maxItems": 20,
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 1,
              "maxLength": 40
            }
          }
        },
        "required": [
          "name",
          "phone"
        ]
      },
...
    }
  },
  "paths": {
    "/clients": {
      "get": {
        "summary": "List authorized clients",
        "tags": [
          "clients"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "range",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "clientId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "provider",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success; JSON {data, meta?}"
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Sign-in required"
          },
          "403": {
            "description": "Role or CSRF denied"
          },
          "404": {
            "description": "Record absent or outside workspace"
          },
          "409": {
            "description": "Duplicate, stale version or invalid transition"
          },
          "422": {
            "description": "Zod validation error with fieldErrors"
          },
          "503": {
            "description": "External integration not configured"
          }
        }
      },
      "post": {
        "summary": "Create clients",
        "tags": [
          "clients"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Client"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success; JSON {data, meta?}"
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Sign-in required"
          },
          "403": {
            "description": "Role or CSRF denied"
          },
          "404": {
            "description": "Record absent or outside workspace"
          },
          "409": {
            "description": "Duplicate, stale version or invalid transition"
          },
          "422": {
            "description": "Zod validation error with fieldErrors"
          },
          "503": {
            "description": "External integration not configured"
          }
        }
      }
    },
...
  }
}